urlShortner/requirements.txt
Django
requests
djangorestframework
djongo==1.3.6